Update based on PR feedback
